third_party/blink/web_tests/wpt_internal/async-script-scheduling/execution-order.html

<!doctype html>
<head>
  <title>Async Script Execution Order</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <link rel="preload" as="script" href="resources/script.js?4&preload">
  <script async blocking=render src="resources/script.js?1&pipe=trickle(d1)"
          data-label="async_blocking_render"></script>
</head>
<body>
  <!-- Note: This test serves mainly to document the current scheduling behavior
       of async scripts with respect to other scripts and events in the loading
       pipeline. It does not represent the "ideal" or "desired" behavior, so
       changing it is probably fine. -->
  <script>
    setup({single_test: true})

    const logs = [];
    function logScript(msg) {
      logs.push(msg);
    }
    window.onload = e => {
      const actualOrder = logs.join(";");
      const expectedOrder = "async_preload;async_blocking_render;async;inline1;external;inline2;defer;DOMContentLoaded";
      assert_equals(actualOrder, expectedOrder);
      done();
    };
    document.addEventListener('DOMContentLoaded', e => { logScript('DOMContentLoaded'); });
  </script>

  <script defer src="resources/script.js?2" data-label="defer"></script>
  <script async src="resources/script.js?3&pipe=trickle(d2)" data-label="async"></script>
  <script async src="resources/script.js?4&preload" data-label="async_preload"></script>
  <link rel=stylesheet href=resources/main.css?pipe=trickle(d3)>
  <script>
    logScript('inline1');
    assert_equals('rgb(255, 255, 0)', getComputedStyle(document.body)['backgroundColor']);
  </script>
  <script src="resources/script.js?5&pipe=trickle(d1)" data-label="external"></script>
  <script>
    logScript('inline2');
  </script>
  <h1>
    This is the first (contentful) paint.
  </h1>
</body>
