third_party/blink/web_tests/platform/generic/virtual/async-script-scheduling-finished-parsing/wpt_internal/async-script-scheduling/execution-order-expected.txt

This is a testharness.js-based test.
FAIL Async Script Execution Order Uncaught Error: assert_equals: expected "async_preload;async_blocking_render;async;inline1;external;inline2;defer;DOMContentLoaded" but got "async_preload;async_blocking_render;inline1;external;inline2;defer;DOMContentLoaded;async"
Harness: the test ran to completion.
